Add SerializedCollections & Photon-Zenject bridge
Add AYellowpaper SerializedCollections plugin (editor and runtime assets) and a Photon-Zenject bridge. Introduce ZenjectNetworkObjectProvider to inject Zenject dependencies into instantiated NetworkObjects. Update domain APIs: rename IColleague.Init to SetMediator and remove CurrentState/PreviousState properties from IStateMachine. Refactor prototype player code: rename movement state classes, replace PlayerMediator with PlayerNetworkEntity, and add many prototype network-entity interfaces and installers.

+++ Assets/ProjectCore/Contexts/Prototype/Features/Proyotype/Player/PlayerMediator/EventPayloads/DamageTakePayload.cs
@@ -2,7 +2,7 @@
 {
     public struct DamageTakePayload
     {
-        public int Amount { get; set; }
+        public byte Amount { get; set; }
         public string DamageType { get; set; }
     }
 }
